+++ react/components/EditorContainer/EditableText/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { IconEdit } from 'vtex.styleguide'
 
 import styles from './styles.css'
 
@@ -20,17 +21,59 @@ const EditableText: React.FC<Props> = ({
   disabled,
   placeholder,
   value,
-}) => (
-  <input
-    className={`bn font-body input-reset ${baseClassName} ${styles.input}`}
-    onChange={onChange}
-    onBlur={onBlur}
-    disabled={disabled}
-    onKeyPress={onKeyPress}
-    placeholder={placeholder || ''}
-    type="text"
-    value={value || ''}
-  />
-)
+}) => {
+  const [isEditing, setIsEditing] = React.useState(false)
+
+  const inputRef = React.useRef<HTMLInputElement | null>(null)
+
+  const startEditing = React.useCallback(() => {
+    inputRef && inputRef.current && inputRef.current.focus()
+  }, [inputRef])
+
+  const handleOnFocus = React.useCallback(() => {
+    setIsEditing(true)
+  }, [setIsEditing])
+
+  const handleOnBlur = React.useCallback(
+    e => {
+      setIsEditing(false)
+      onBlur && onBlur(e)
+    },
+    [setIsEditing, onBlur]
+  )
+
+  return (
+    <div
+      role="presentation"
+      className={`flex items-baseline bb b--light-gray ${
+        isEditing ? 'hover-b--action-primary' : ''
+      } pb2 ${isEditing ? 'b--action-primary' : ''} ${
+        styles.editableTextInputWrapper
+      }`}
+      style={{ cursor: 'text' }}
+      onClick={startEditing}
+    >
+      <input
+        className={`flex-grow-1 w-100 truncate outline-0 c-on-base bn font-body input-reset ${baseClassName} ${styles.input}`}
+        onChange={onChange}
+        onFocus={handleOnFocus}
+        onBlur={handleOnBlur}
+        disabled={disabled}
+        onKeyPress={onKeyPress}
+        placeholder={placeholder || ''}
+        ref={inputRef}
+        type="text"
+        value={value || ''}
+      />
+      <span
+        className={`${styles.editableTextInputWrapper__input} ${
+          isEditing ? 'dn' : ''
+        }`}
+      >
+        <IconEdit />
+      </span>
+    </div>
+  )
+}
 
 export default EditableText